+++ commands/initInitialisationCommands.go
@@ -1,12 +1,13 @@
 package commands
 
 import (
+	"github.com/ditrit/badaas/configuration"
 	"github.com/ditrit/verdeter"
 )
 
 func initInitialisationCommands(cfg *verdeter.VerdeterCommand) {
 
-	cfg.GKey("default.admin.password", verdeter.IsStr, "",
+	cfg.GKey(configuration.InitializationDefaultAdminPasswordKey, verdeter.IsStr, "",
 		"Set the default admin password is the admin user is not created yet.")
-	cfg.SetDefault("default.admin.password", "admin")
+	cfg.SetDefault(configuration.InitializationDefaultAdminPasswordKey, "admin")
 }