+++ commands/initLoggerCommands.go
@@ -1,18 +1,16 @@
 package commands
 
 import (
+	"github.com/ditrit/badaas/configuration"
 	"github.com/ditrit/verdeter"
 	"github.com/ditrit/verdeter/validators"
 )
 
 func initLoggerCommands(cfg *verdeter.VerdeterCommand) {
+	cfg.GKey(configuration.LoggerModeKey, verdeter.IsStr, "", "The logger mode (default to \"prod\")")
+	cfg.SetDefault(configuration.LoggerModeKey, "prod")
+	cfg.AddValidator(configuration.LoggerModeKey, validators.AuthorizedValues("authorized values", "prod", "dev"))
 
-	loggerModeKey := "logger.mode"
-	cfg.GKey(loggerModeKey, verdeter.IsStr, "", "The logger mode (default to \"prod\")")
-	cfg.SetDefault(loggerModeKey, "prod")
-	cfg.AddValidator(loggerModeKey, validators.AuthorizedValues("authorized values", "prod", "dev"))
-
-	loggerRequestTemplateKey := "logger.request.template"
-	cfg.GKey(loggerRequestTemplateKey, verdeter.IsStr, "", "Template message for all request logs")
-	cfg.SetDefault(loggerRequestTemplateKey, "Receive {{method}} request on {{url}}")
+	cfg.GKey(configuration.LoggerRequestTemplateKey, verdeter.IsStr, "", "Template message for all request logs")
+	cfg.SetDefault(configuration.LoggerRequestTemplateKey, "Receive {{method}} request on {{url}}")
 }