+++ changelog.md
@@ -27,5 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a CONTRIBUTING.md and a documentation file for configuration (configuration.md)
 - Add a session services.
 - Add a basic authentification controller.
+- Now config keys are only declared once with constants in the `configuration/` package.
 
 [unreleased]: https://github.com/ditrit/badaas/blob/main/changelog.md#unreleased